requirements.txt
# Core utilities
tqdm
numpy
pandas
PyYAML

# Academic papers ingestion
arxiv
pymupdf        # (imported as "fitz")

# NLP / Embeddings / Retrieval
sentence-transformers
faiss-cpu
rank-bm25
ollama
pyspellchecker
keybert
sentence-transformers
pymupdf 
pdfplumber 
open-clip-torch 
pillow 
tqdm

# Reranking (cross-encoders)
transformers

# Evaluation
ragas
datasets        # needed for ragas + huggingface datasets

# Web demo UI
streamlit

# Optional image processing for multi-modal PDF
pytesseract
pillow